+++ TEP_project/KnapsackProblem.h
@@ -5,38 +5,36 @@
 
 class KnapsackProblem
 {
-	size_t _itemCount;
-	size_t _maxWeight;
+	int _itemCount;
+	int _maxWeight;
 	int* _itemWeights;
 	int* _itemValues;
 
 	int _bestFit;
 	bool* _bestSolution;
 
-
 public:
-
 	KnapsackProblem();
 
 	KnapsackProblem(const KnapsackProblem& other);
 	
 	KnapsackProblem(KnapsackProblem&& other) noexcept;
 	
-	KnapsackProblem(const size_t& itemCount, const size_t& maxWeight, int* itemWeights, int* itemValues);
+	KnapsackProblem(const int& itemCount, const int& maxWeight, int* itemWeights, int* itemValues);
 
-	void LoadProblem(const size_t& itemCount, const size_t& maxWeight, int* itemWeights, int* itemValues);
+	void LoadProblem(const int& itemCount, const int& maxWeight, int* itemWeights, int* itemValues);
 	
 	KnapsackProblem operator=(const KnapsackProblem& other);
 
 	KnapsackProblem& operator=(KnapsackProblem&& other) noexcept;
 
-	int CalculateFitness(const bool* solution, const size_t& solSize);
+	int CalculateFitness(const bool* solution, const int& solSize);
 
 	~KnapsackProblem();
 
-	size_t GetItemCount() const;
+	int GetItemCount() const;
 	
-	size_t GetMaxWeight() const;
+	int GetMaxWeight() const;
 
 	void PrintBestFit();
 
@@ -46,15 +44,15 @@ public:
 private:
 
 	template<typename T>
-	void PrintArray(const size_t& size, const T* arr);
+	void PrintArray(const int& size, const T* arr);
 
-	void ArrayCopy(bool*& container, const size_t& size, const bool* copyFrom);
+	void ArrayCopy(bool*& container, const int& size, const bool* copyFrom);
 };
 
 template<typename T>
-inline void KnapsackProblem::PrintArray(const size_t& size, const T* arr)
+inline void KnapsackProblem::PrintArray(const int& size, const T* arr)
 {
-	for (size_t i = 0; i < size; ++i)
+	for (int i = 0; i < size; ++i)
 		std::cout << arr[i] << " ";
 
 	std::cout << std::endl;
